+++ NEWS.md
@@ -7,6 +7,8 @@ Version 1.4.6
 * Improved scaling for ordiplot with quadratic model and with biplot = FALSE
 * optima.gllvm and tolerances.gllvm for num.lv now correctly provide tolerances w.r.t. the scaled LV
 * Improved starting values for models with 'randomB'
+* 'which.Xcoef' in coefplot.gllvm now also works for fourth-corner models
+* Added intercept if beta0com=TRUE to coefplot.gllvm for fourth-corner models
 
 ## Bug Fixes
 * Bug fixed that prevented increasing he point size of sites in ordiplot with symbols = TRUE